refactored extension impls and added javadoc to embeddedHiveMQ builder.
